requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
email-validator
pandas
pytest
python-dotenv
gunicorn
pyngrok